% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapowerAD.R
\docType{data}
\name{data.powerAD}
\alias{data.powerAD}
\title{data.powerAD}
\format{
An object of class \code{data.frame} with 346500 rows and 6 columns.
}
\usage{
data(data.powerAD)
}
\description{
Dataset containing all results from the power estimations in Pirla, Taquet and Quoidbach (2021).
The data is structured in a dataframe and organized based on power, sampling, effect size, alpha level (p.value), and metric of interest.
A detailed description of the estimation procedures is included in the reference paper.
}
\examples{
data(data.powerAD)
results <- data.powerAD
}
\references{
Pirla, Taquet and Quoidbach (2021). ADD REFERENCE
}
\keyword{Affect}
\keyword{Analysis}
\keyword{Datasets,}
\keyword{Dynamics,}
\keyword{Power}
