% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsizes.R
\docType{data}
\name{effect.sizes}
\alias{effect.sizes}
\title{effect.sizes}
\format{
An object of class \code{data.frame} with 22 rows and 8 columns.
}
\usage{
data(effect.sizes)
}
\description{
Dataset containing the benchmark plausible effect sizes for affect dynamics studies presented in Pirla, Taquet and Quoidbach (2021).
}
\examples{
data(effect.sizes)
benchmarks <- effect.sizes
}
\references{
Pirla, Taquet and Quoidbach (2021). ADD REFERENCE
}
\keyword{Affect}
\keyword{Analysis}
\keyword{Datasets,}
\keyword{Dynamics,}
\keyword{Power}
