% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerAD.R
\name{powerAD}
\alias{powerAD}
\title{powerAD}
\usage{
powerAD(individuals, samples, metric, r, p.value = 0.05)
}
\arguments{
\item{individuals}{Number of individuals sampled.}

\item{samples}{Number of affect observations per individual.}

\item{metric}{Metric of interest. Must be a single character from "Average", "Rel.SD", "SD", "RMSSD", "TKEO", "PAC" or "Autocorrelation".}

\item{r}{Number from 0.01 to 0.99 indicating the expected effect size (Pearson correlation) of interest.}

\item{p.value}{Alpha level. Must be one of the following numbers: 0.01, 0.05, 0.001, 0.005, 0.001. Defaults to 0.05.}
}
\description{
Estimates the statistical power of an affect dynamics time series study for a given sampling, effect size, alpha level and metric of interest. Based on Pirla, Taquet and Quoidbach (2021).
}
\examples{
powerAD(individuals=500, samples=10, metric="SD", r=0.1, p.value=0.01)
}
\references{
Pirla, Taquet and Quoidbach (2021). ADD REFERENCE
}
\keyword{Affect}
\keyword{Analysis}
\keyword{Dynamics,}
\keyword{Power}
